moved readSbol function into sbolservice
