+++ backend/src/main/java/com/synbiohub/sbh3/dto/SubmitFileFormat.java
@@ -1,0 +1,5 @@
+package com.synbiohub.sbh3.dto;
+
+public enum SubmitFileFormat {
+    SBOL, GENBANK, FASTA, GFF3, ATTACHMENT
+}
