+++ backend/src/main/java/com/synbiohub/sbh3/services/SubmitServiceImpl.java
@@ -1,14 +1,15 @@
 package com.synbiohub.sbh3.services;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.synbiohub.sbh3.security.model.AuthCodes;
 import com.synbiohub.sbh3.security.repo.AuthRepository;
 import com.synbiohub.sbh3.sparql.SPARQLQuery;
 import com.synbiohub.sbh3.submit.SubmitPayload;
 import com.synbiohub.sbh3.submit.SubmitPluginService;
 import com.synbiohub.sbh3.utils.ConfigUtil;
+import com.synbiohub.sbh3.utils.FileUtil;
+import com.synbiohub.sbh3.utils.StringUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FilenameUtils;
@@ -20,34 +21,27 @@ import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 import org.apache.hc.client5.http.impl.classic.HttpClients;
 import org.apache.hc.core5.http.ContentType;
 import org.apache.hc.core5.http.io.entity.ByteArrayEntity;
-import org.sbolstandard.core2.*;
-import org.springframework.http.*;
+import org.sbolstandard.core2.SBOLValidationException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
-import org.synbiohub.frontend.SynBioHubException;
-import org.synbiohub.frontend.SynBioHubFrontend;
 
-import javax.xml.namespace.QName;
-import java.io.*;
-import java.net.URI;
-import java.net.URISyntaxException;
+import java.io.IOException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
-import java.util.stream.Collectors;
 import java.util.zip.GZIPOutputStream;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
 
 /**
  * Orchestrates {@code POST /submit}: multipart form → Virtuoso graph upload.
@@ -70,6 +64,8 @@ public class SubmitServiceImpl implements SubmitService {
     private final CitationService citationService;
 
     private final CollectionService collectionService;
+    private final SbolService sbolService;
+    private final ObjectMapper objectMapper;
 
     /**
      * Main submit entry point. Each step mutates {@code payload} in place.
@@ -79,7 +75,9 @@ public class SubmitServiceImpl implements SubmitService {
         SubmitPayload payload = parse(allParams, file);
         collectionService.sanitize(payload);
         submitPluginService.applySubmitPlugin(payload); // optional transform of uploaded file
-        readSbol(payload);
+        sbolService.readSbol(payload, webOfRegistriesMap(),
+                ConfigUtil.get("databasePrefix").asText() + "user/" + payload.getCreatedBy(),
+                resolveUserAuthToken(payload));
         prepare(payload);  // only runs for overwrite_merge == 1
         upload(payload);
         return successResponse();
@@ -91,7 +89,7 @@ public class SubmitServiceImpl implements SubmitService {
 
     /**
      * Reads form parameters and persists the uploaded file to a temp path.
-     * Does not validate business rules (that is {@link #sanitize}).
+     * Does not validate business rules.
      */
     private SubmitPayload parse(SubmitPayload payload, MultipartFile file) throws IOException {
         //TODO: why are we creating a temp file here?
@@ -108,420 +106,6 @@ public class SubmitServiceImpl implements SubmitService {
         return payload;
     }
 
-    // -------------------------------------------------------------------------
-    // readSbol — validate/convert input, build composite SBOLDocument (PrepareSubmissionJob)
-    // -------------------------------------------------------------------------
-
-    /**
-     * SBOL preparation pipeline. Produces {@code payload.resultFilePath} (serialized XML).
-     */
-    private void readSbol(SubmitPayload payload) throws IOException, SBOLValidationException {
-        setupReadSbol(payload);
-        classifySubmitInput(payload);
-        mergeValidatedSbolFiles(payload);
-        prepareRootCollection(payload);
-        mergeIntoExistingCollection(payload);   // modes 2/3 only
-        fixMutableAnnotations(payload);
-        populateCollectionMembership(payload);
-        if (ConfigUtil.get("useSBOLExplorer").asBoolean(false)) {
-            incrementallyUpdateSbolExplorer(payload);
-        }
-        finishReadSbol(payload);
-    }
-
-    /** Writes the composite document to a temp XML file ({@code resultFilename} in legacy Node). */
-    private void finishReadSbol(SubmitPayload payload) throws IOException {
-        Path resultFile = Files.createTempFile("sbh_convert_validate", ".xml");
-        try {
-            SBOLWriter.write(payload.getSbolDocument(), resultFile.toFile());
-        } catch (SBOLConversionException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    "Failed to write SBOL output: " + e.getMessage());
-        }
-        payload.setResultFilePath(resultFile.toAbsolutePath().toString());
-
-        long started = payload.getReadSbolStartedAtMs() != null
-                ? payload.getReadSbolStartedAtMs()
-                : System.currentTimeMillis();
-        log.info("readSbol total time (sec): {}", (System.currentTimeMillis() - started) / 1000.0);
-    }
-
-    /** Notifies SBOL Explorer of added top-levels after a successful submit (when enabled). */
-    private void incrementallyUpdateSbolExplorer(SubmitPayload payload) throws IOException {
-        String graph = resolveOwnedByPrefix(payload);
-
-        ObjectNode body = MAPPER.createObjectNode();
-        body.set("partsToRemove", MAPPER.createArrayNode());
-        ArrayNode partsToAdd = MAPPER.createArrayNode();
-        for (TopLevel topLevel : payload.getSbolDocument().getTopLevels()) {
-            ObjectNode part = MAPPER.createObjectNode();
-            part.put("subject", topLevel.getIdentity().toString());
-            part.put("displayId", topLevel.getDisplayId());
-            part.put("version", topLevel.getVersion());
-            part.put("name", topLevel.getName());
-            part.put("description", topLevel.getDescription());
-            part.put("type", "TODO");
-            part.put("graph", graph);
-            partsToAdd.add(part);
-        }
-        body.set("partsToAdd", partsToAdd);
-
-        String endpoint = ConfigUtil.get("SBOLExplorerEndpoint").asText();
-        if (!endpoint.endsWith("/")) {
-            endpoint += "/";
-        }
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        try {
-            new RestTemplate().postForEntity(endpoint + "incrementalupdate",
-                    new HttpEntity<>(body, headers), String.class);
-        } catch (Exception e) {
-            log.warn("SBOLExplorer /incrementalupdate failed", e);
-        }
-    }
-
-    /**
-     * Validates each SBOL input file, strips registry objects, rewrites URIs to the submission
-     * prefix, and merges into the composite document on {@code payload}.
-     */
-    private void mergeValidatedSbolFiles(SubmitPayload payload) throws IOException {
-        String databasePrefix = ConfigUtil.get("databasePrefix").asText();
-        String uriPrefix = collectionService.uriPrefix(payload);
-        String version = payload.getVersion() != null ? payload.getVersion() : "1";
-
-        boolean requireComplete = ConfigUtil.get("requireComplete").asBoolean(false);
-        boolean requireCompliant = ConfigUtil.get("requireCompliant").asBoolean(false);
-        boolean enforceBestPractices = ConfigUtil.get("requireBestPractice").asBoolean(false);
-        List<String> registryPrefixes = new ArrayList<>(webOfRegistriesMap().keySet());
-
-        StringBuilder readLog = new StringBuilder();
-        SBOLDocument composite = payload.getSbolDocument();
-
-        for (Path file : payload.getSbolFiles()) {
-            String filename = file.toString();
-            String defaultDisplayId = fixDisplayId(displayIdFromFilename(filename));
-
-            ByteArrayOutputStream logOut = new ByteArrayOutputStream();
-            ByteArrayOutputStream errorOut = new ByteArrayOutputStream();
-            SBOLDocument individual = SBOLValidate.validate(
-                    new PrintStream(logOut),
-                    new PrintStream(errorOut),
-                    filename,
-                    "http://dummy.org/",
-                    defaultDisplayId,
-                    requireComplete,
-                    requireCompliant,
-                    enforceBestPractices,
-                    false,
-                    "1",
-                    true,
-                    "",
-                    "",
-                    filename,
-                    "",
-                    false,
-                    false,
-                    false,
-                    false,
-                    false,
-                    null,
-                    false,
-                    true,
-                    false);
-
-            readLog.append("[").append(filename).append(" log] \n")
-                    .append(logOut.toString(StandardCharsets.UTF_8)).append("\n");
-
-            String errorLog = errorOut.toString(StandardCharsets.UTF_8);
-            if (errorLog.startsWith("File is empty")) {
-                // Empty Office/plugin placeholder — treat as blank document.
-                individual = new SBOLDocument();
-                errorLog = "";
-            } else if (!errorLog.isEmpty()) {
-                failReadSbol(readLog.toString(), errorLog);
-            }
-
-            // Registry URIs are removed from the document but remembered for collection membership.
-            stripRegistryTopLevels(individual, registryPrefixes, payload);
-
-            try {
-                individual = rewriteIndividualUris(individual, uriPrefix, version, databasePrefix);
-                composite.createCopy(individual);
-            } catch (Exception e) {
-                StringWriter sw = new StringWriter();
-                e.printStackTrace(new PrintWriter(sw));
-                failReadSbol(readLog.toString(), sw.toString());
-            }
-        }
-
-        if (!readLog.isEmpty()) {
-            log.debug("readSbol validate log:\n{}", readLog);
-        }
-    }
-
-    /**
-     * Creates or updates the root {@code Collection} and annotates every top-level with
-     * ownedBy, topLevel, citations, and creator metadata.
-     */
-    private void prepareRootCollection(SubmitPayload payload) throws IOException, SBOLValidationException {
-        SBOLDocument doc = payload.getSbolDocument();
-        String version = payload.getVersion() != null ? payload.getVersion() : "1";
-        String ownedByUri = resolveOwnedByPrefix(payload);
-
-        String displayId = payload.collectionDisplayId();
-        if (displayId == null) {
-            return;
-        }
-        org.sbolstandard.core2.Collection root = doc.getCollection(displayId, version);
-        if (root == null) {
-            root = doc.createCollection(displayId, version);
-            log.debug("New root collection: {}", root.getIdentity());
-        }
-        applySubmitRootCollection(payload, doc, root, ownedByUri);
-    }
-
-    /** Sets name, description, creator, citations, ownedBy, and topLevel on all identified objects. */
-    private void applySubmitRootCollection(SubmitPayload payload, SBOLDocument doc,
-                                           org.sbolstandard.core2.Collection root,
-                                           String ownedByUri) {
-        try {
-            String creator = userService.getUserByUsername(payload.getCreatedBy()).getName();
-            if (creator != null) {
-                root.createAnnotation(DC_CREATOR, creator);
-            }
-            root.createAnnotation(DCTERMS_CREATED,
-                    ZonedDateTime.now().format(DateTimeFormatter.ISO_INSTANT));
-            if (payload.getName() != null) {
-                root.setName(payload.getName());
-            }
-            if (payload.getDescription() != null) {
-                root.setDescription(payload.getDescription());
-            }
-            new IdentifiedVisitor() {
-                @Override
-                public void visit(Identified identified, TopLevel topLevel) {
-                    try {
-                        propagateTopLevelInNestedAnnotations(topLevel, identified.getAnnotations());
-                        for (Integer pubmedId : payload.getCitationPubmedIds()) {
-                            identified.createAnnotation(OBO_PUBMED, String.valueOf(pubmedId));
-                        }
-                        Annotation ownedBy = identified.getAnnotation(SBH_OWNED_BY);
-                        if (ownedBy != null) {
-                            identified.removeAnnotation(ownedBy);
-                        }
-                        identified.createAnnotation(SBH_OWNED_BY, new URI(ownedByUri));
-                        Annotation topLevelAnn = identified.getAnnotation(SBH_TOP_LEVEL);
-                        if (topLevelAnn != null) {
-                            identified.removeAnnotation(topLevelAnn);
-                        }
-                        identified.createAnnotation(SBH_TOP_LEVEL, topLevel.getIdentity());
-                    } catch (SBOLValidationException | URISyntaxException e) {
-                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                                "Failed to annotate object: " + e.getMessage());
-                    }
-                }
-            }.visitDocument(doc);
-        } catch (SBOLValidationException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    "Failed to prepare root collection: " + e.getMessage());
-        }
-    }
-
-    /**
-     * Merge modes 2/3: for objects whose URIs already exist in a configured registry,
-     * fetch from remote SynBioHub and either reject (different content), replace (mode 3),
-     * or keep as collection members (identical content).
-     */
-    private void mergeIntoExistingCollection(SubmitPayload payload) throws IOException {
-        String mode = payload.getOverwriteMerge();
-        if (mode == null || "0".equals(mode) || "1".equals(mode)) {
-            return;
-        }
-
-        SBOLDocument doc = payload.getSbolDocument();
-        String rootCollectionIdentity = payload.getCollectionUri();
-        String version = payload.getVersion() != null ? payload.getVersion() : "1";
-        org.sbolstandard.core2.Collection rootCollection = null;
-        if (payload.collectionDisplayId() != null) {
-            rootCollection = doc.getCollection(payload.collectionDisplayId(), version);
-        }
-
-        String shareLinkSalt = ConfigUtil.get("shareLinkSalt").asText("");
-        String authToken = resolveUserAuthToken(payload);
-        Map<String, String> registries = webOfRegistriesMap();
-
-        for (TopLevel topLevel : new ArrayList<>(doc.getTopLevels())) {
-            String identity = topLevel.getIdentity().toString();
-
-            if (rootCollectionIdentity != null && identity.equals(rootCollectionIdentity)) {
-                // Existing collection metadata is preserved from the store during merge.
-                stripExistingRootCollectionMetadata(topLevel);
-                continue;
-            }
-
-            for (Map.Entry<String, String> entry : registries.entrySet()) {
-                String registry = entry.getKey();
-                if (!identity.startsWith(registry)) {
-                    continue;
-                }
-
-                String fetchUri = registryFetchUri(identity, registry, shareLinkSalt);
-                SynBioHubFrontend sbh = new SynBioHubFrontend(entry.getValue(), registry);
-                if (authToken != null) {
-                    sbh.setUser(authToken);
-                }
-
-                SBOLDocument remoteDoc;
-                try {
-                    remoteDoc = sbh.getSBOL(URI.create(fetchUri));
-                } catch (SynBioHubException e) {
-                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
-                }
-
-                if (remoteDoc == null) {
-                    break;
-                }
-
-                TopLevel remote = remoteDoc.getTopLevel(topLevel.getIdentity());
-                if (remote == null) {
-                    break;
-                }
-
-                if (!topLevel.equals(remote)) {
-                    if ("3".equals(mode)) {
-                        try {
-                            sbh.replaceSBOL(URI.create(fetchUri));
-                        } catch (SynBioHubException e) {
-                            log.warn("replaceSBOL failed for {}", identity, e);
-                        }
-                    } else {
-                        throw new ResponseStatusException(HttpStatus.CONFLICT,
-                                "Submission terminated.\nA submission with this id already exists,"
-                                        + " and it includes an object: " + identity
-                                        + " that is already in this repository and has different content");
-                    }
-                } else {
-                    log.debug("Found duplicate registry object, keeping as member: {}", identity);
-                    try {
-                        doc.removeTopLevel(topLevel);
-                        if (rootCollection != null) {
-                            rootCollection.addMember(topLevel.getIdentity());
-                        }
-                    } catch (SBOLValidationException e) {
-                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                                "Failed to merge duplicate object: " + e.getMessage());
-                    }
-                }
-                break;
-            }
-        }
-    }
-
-    /** Clears name/description/creator on the existing root collection so merge keeps store metadata. */
-    private static void stripExistingRootCollectionMetadata(TopLevel topLevel) {
-        topLevel.unsetDescription();
-        topLevel.unsetName();
-        topLevel.clearWasDerivedFroms();
-        Annotation creator = topLevel.getAnnotation(DC_CREATOR);
-        if (creator != null) {
-            topLevel.removeAnnotation(creator);
-        }
-    }
-
-    /** Rewrites user-scoped image paths in sbh:mutable* HTML to public collection URLs. */
-    private void fixMutableAnnotations(SubmitPayload payload) {
-        String displayId = payload.collectionDisplayId();
-        if (displayId == null) {
-            return;
-        }
-        String publicPrefix = "img src=\"/public/" + displayId.replace("_collection", "") + "/";
-        SBOLDocument doc = payload.getSbolDocument();
-        for (TopLevel topLevel : doc.getTopLevels()) {
-            rewriteMutableAnnotation(topLevel, SBH_MUTABLE_DESCRIPTION, publicPrefix);
-            rewriteMutableAnnotation(topLevel, SBH_MUTABLE_NOTES, publicPrefix);
-            rewriteMutableAnnotation(topLevel, SBH_MUTABLE_PROVENANCE, publicPrefix);
-        }
-    }
-
-    private void rewriteMutableAnnotation(TopLevel topLevel, QName qname, String publicPrefix) {
-        Annotation annotation = topLevel.getAnnotation(qname);
-        if (annotation == null || !annotation.isStringValue()) {
-            return;
-        }
-        String value = annotation.getStringValue();
-        String updated = MUTABLE_IMG_USER_PATH.matcher(value).replaceAll(publicPrefix);
-        if (value.equals(updated)) {
-            return;
-        }
-        try {
-            topLevel.removeAnnotation(annotation);
-            topLevel.createAnnotation(qname, updated);
-        } catch (SBOLValidationException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    "Failed to rewrite mutable annotation: " + e.getMessage());
-        }
-    }
-
-    /**
-     * Adds all top-levels as collection members. Re-adds registry objects stripped earlier.
-     */
-    private void populateCollectionMembership(SubmitPayload payload) {
-        String displayId = payload.collectionDisplayId();
-        if (displayId == null) {
-            return;
-        }
-        String version = payload.getVersion() != null ? payload.getVersion() : "1";
-        SBOLDocument doc = payload.getSbolDocument();
-        org.sbolstandard.core2.Collection rootCollection = doc.getCollection(displayId, version);
-        if (rootCollection == null) {
-            return;
-        }
-
-        URI rootIdentity = rootCollection.getIdentity();
-        try {
-            for (TopLevel topLevel : doc.getTopLevels()) {
-                if (rootIdentity.equals(topLevel.getIdentity())) {
-                    continue;
-                }
-                rootCollection.addMember(topLevel.getIdentity());
-            }
-            for (String uri : payload.getUrisFoundInSynBioHub()) {
-                rootCollection.addMember(URI.create(uri));
-            }
-        } catch (SBOLValidationException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    "Failed to populate collection membership: " + e.getMessage());
-        }
-    }
-
-    /** Share token for fetching private user objects from remote registries during merge. */
-    private static String registryFetchUri(String identity, String registry, String shareLinkSalt) {
-        if (identity.startsWith(registry + "/user/")) {
-            return identity + "/" + privateShareToken(identity, shareLinkSalt) + "/share";
-        }
-        return identity;
-    }
-
-    private static String privateShareToken(String identity, String shareLinkSalt) {
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-1");
-            String innerHex = sha1Hex(md.digest((identity + "/edit").getBytes(StandardCharsets.UTF_8)));
-            md.reset();
-            String payload = "synbiohub_" + innerHex + (shareLinkSalt == null ? "" : shareLinkSalt);
-            return sha1Hex(md.digest(payload.getBytes(StandardCharsets.UTF_8)));
-        } catch (NoSuchAlgorithmException e) {
-            throw new IllegalStateException("SHA-1 not available", e);
-        }
-    }
-
-    private static String sha1Hex(byte[] digest) {
-        StringBuilder sb = new StringBuilder(digest.length * 2);
-        for (byte b : digest) {
-            sb.append(String.format(Locale.US, "%02x", b & 0xff));
-        }
-        return sb.toString();
-    }
-
     private String resolveUserAuthToken(SubmitPayload payload) {
         return authRepository.findByName(payload.getCreatedBy())
                 .map(AuthCodes::getAuth)
@@ -535,252 +119,6 @@ public class SubmitServiceImpl implements SubmitService {
             node.fields().forEachRemaining(entry -> map.put(entry.getKey(), entry.getValue().asText()));
         }
         return map;
-    }
-
-    /** Ensures sbh:topLevel is set on nested annotation trees (e.g. mutable HTML blocks). */
-    private static void propagateTopLevelInNestedAnnotations(TopLevel topLevel, List<Annotation> annotations)
-            throws SBOLValidationException {
-        for (Annotation annotation : annotations) {
-            if (!annotation.isNestedAnnotations()) {
-                continue;
-            }
-            List<Annotation> nested = new ArrayList<>(annotation.getAnnotations());
-            propagateTopLevelInNestedAnnotations(topLevel, nested);
-            nested.removeIf(a -> SBH_TOP_LEVEL.equals(a.getQName()));
-            nested.add(new Annotation(SBH_TOP_LEVEL, topLevel.getIdentity()));
-            annotation.setAnnotations(nested);
-        }
-    }
-
-    /**
-     * Removes top-levels that belong to a configured web-of-registries prefix.
-     * Their URIs are recorded so they can be re-added as collection members later.
-     */
-    private static void stripRegistryTopLevels(SBOLDocument individual,
-                                                 List<String> registryPrefixes,
-                                                 SubmitPayload payload) {
-        for (TopLevel topLevel : new ArrayList<>(individual.getTopLevels())) {
-            String identity = topLevel.getIdentity().toString();
-            for (String registry : registryPrefixes) {
-                if (identity.startsWith(registry)) {
-                    log.debug("Found and removed registry object: {}", identity);
-                    payload.getUrisFoundInSynBioHub().add(identity);
-                    try {
-                        individual.removeTopLevel(topLevel);
-                    } catch (SBOLValidationException e) {
-                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                                "Failed to remove registry object: " + e.getMessage());
-                    }
-                    break;
-                }
-            }
-        }
-    }
-
-    /** Assigns submission URI prefix/version and fixes attachment download sources after rewrite. */
-    private static SBOLDocument rewriteIndividualUris(SBOLDocument individual, String uriPrefix, String version,
-                                                    String databasePrefix) throws SBOLValidationException {
-        individual.setDefaultURIprefix("http://dummy.org/");
-        if (uriPrefix == null) {
-            return individual;
-        }
-        if (individual.getTopLevels().isEmpty()) {
-            individual.setDefaultURIprefix(uriPrefix);
-            return individual;
-        }
-        individual = individual.changeURIPrefixVersion(uriPrefix, null, version);
-        individual.setDefaultURIprefix(uriPrefix);
-        for (Attachment attachment : individual.getAttachments()) {
-            String source = attachment.getSource().toString();
-            if (source.startsWith(databasePrefix) && source.endsWith("/download")) {
-                attachment.setSource(URI.create(attachment.getIdentity().toString() + "/download"));
-            }
-        }
-        return individual;
-    }
-
-    private static String displayIdFromFilename(String filename) {
-        String displayId = filename;
-        int dot = displayId.lastIndexOf('.');
-        if (dot != -1) {
-            displayId = displayId.substring(0, dot);
-        }
-        int slash = displayId.lastIndexOf('/');
-        if (slash != -1) {
-            displayId = displayId.substring(slash + 1);
-        }
-        return displayId;
-    }
-
-    private static String fixDisplayId(String displayId) {
-        if (displayId == null || displayId.isEmpty()) {
-            return "_";
-        }
-        displayId = displayId.replaceAll("[^a-zA-Z0-9_]", "_");
-        displayId = displayId.replace(" ", "_");
-        if (Character.isDigit(displayId.charAt(0))) {
-            displayId = "_" + displayId;
-        }
-        return displayId;
-    }
-
-    /** {@code sbh:user/<username>} URI used for ownedBy annotations. */
-    private static String resolveOwnedByPrefix(SubmitPayload payload) throws IOException {
-        return ConfigUtil.get("databasePrefix").asText() + "user/" + payload.getCreatedBy();
-    }
-
-    private void failReadSbol(String readLog, String errorLog) {
-        log.debug("readSbol failed:\n{}\n{}", readLog, errorLog);
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, errorLog);
-    }
-
-    // --- readSbol: classify uploaded file(s) into sbolFiles / attachmentFiles ---
-
-    /** Unpacks or classifies the uploaded file. */
-    private void classifySubmitInput(SubmitPayload payload) throws IOException {
-        if (payload.getUploadedFilePath() == null) {
-            return;
-        }
-        Path input = Path.of(payload.getUploadedFilePath());
-        if (!Files.isRegularFile(input)) {
-            return;
-        }
-
-        String ext = fileExtension(input);
-        if (OFFICE_EXTENSIONS.contains(ext)) {
-            payload.getSbolFiles().add(input);
-            return;
-        }
-        if ("omex".equals(ext) || isCombineArchive(input)) {
-            extractSubmitArchive(input, payload, true);
-            return;
-        }
-        if ("zip".equals(ext)) {
-            // Generic zip: unpack and classify; not flagged as COMBINE.
-            extractSubmitArchive(input, payload, false);
-            return;
-        }
-
-        classifySubmitFile(input, payload);
-    }
-
-    /** Extracts a zip/omex archive and classifies each entry; sets {@code extractDirPath}. */
-    private void extractSubmitArchive(Path archive, SubmitPayload payload, boolean combine)
-            throws IOException {
-        Path dest = Files.createTempDirectory("sbh-submit-unpack-");
-        payload.setExtractDirPath(dest.toAbsolutePath().toString());
-        try (ZipFile zip = new ZipFile(archive.toFile())) {
-            var entries = zip.entries();
-            while (entries.hasMoreElements()) {
-                ZipEntry entry = entries.nextElement();
-                if (entry.isDirectory()) {
-                    continue;
-                }
-                Path out = dest.resolve(entry.getName()).normalize();
-                if (!out.startsWith(dest)) {
-                    continue; // zip-slip guard
-                }
-                Files.createDirectories(out.getParent());
-                try (InputStream in = zip.getInputStream(entry)) {
-                    Files.copy(in, out);
-                }
-                classifySubmitFile(out, payload);
-            }
-        }
-        if (combine && payload.getSbolFiles().isEmpty()) {
-            payload.getSbolFiles().add(archive);
-        }
-    }
-
-    /** COMBINE archives contain a manifest.xml entry. */
-    private boolean isCombineArchive(Path file) throws IOException {
-        if (!"zip".equals(fileExtension(file)) && !"omex".equals(fileExtension(file))) {
-            return false;
-        }
-        try (ZipFile zip = new ZipFile(file.toFile())) {
-            return zip.getEntry("manifest.xml") != null
-                    || zip.stream().anyMatch(e -> e.getName().endsWith("manifest.xml"));
-        }
-    }
-
-    /** Routes a single file into sbolFiles or attachmentFiles. */
-    private void classifySubmitFile(Path file, SubmitPayload payload) throws IOException {
-        switch (guessSubmitFileFormat(file)) {
-            case SBOL, GENBANK, FASTA, GFF3 -> payload.getSbolFiles().add(file);
-            case ATTACHMENT -> payload.getAttachmentFiles().add(file);
-        }
-    }
-
-    private enum SubmitFileFormat {
-        SBOL, GENBANK, FASTA, GFF3, ATTACHMENT
-    }
-
-    /** Sniffs file content and extension; anything unrecognized becomes an attachment. */
-    private SubmitFileFormat guessSubmitFileFormat(Path file) throws IOException {
-        String path = file.toAbsolutePath().toString();
-        if (SBOLReader.isGenBankFile(path)) {
-            return SubmitFileFormat.GENBANK;
-        }
-        if (SBOLReader.isFastaFile(path)) {
-            return SubmitFileFormat.FASTA;
-        }
-        if (SBOLReader.isGFF3File(path)) {
-            return SubmitFileFormat.GFF3;
-        }
-
-        String head = readFileHead(file, 4096).toLowerCase();
-        if (head.contains("sbols.org") || head.contains("sbol.org") || head.contains("<rdf:rdf")) {
-            return SubmitFileFormat.SBOL;
-        }
-
-        String ext = fileExtension(file);
-        if (Set.of("xml", "sbol", "rdf").contains(ext)) {
-            return SubmitFileFormat.SBOL;
-        }
-        if ("gb".equals(ext) || "gbk".equals(ext)) {
-            return SubmitFileFormat.GENBANK;
-        }
-        if ("fasta".equals(ext) || "fa".equals(ext)) {
-            return SubmitFileFormat.FASTA;
-        }
-        if ("gff".equals(ext) || "gff3".equals(ext)) {
-            return SubmitFileFormat.GFF3;
-        }
-
-        return SubmitFileFormat.ATTACHMENT;
-    }
-
-    private static String readFileHead(Path file, int maxBytes) throws IOException {
-        byte[] buf = new byte[maxBytes];
-        int read;
-        try (InputStream in = Files.newInputStream(file)) {
-            read = in.read(buf);
-        }
-        if (read <= 0) {
-            return "";
-        }
-        return new String(buf, 0, read, StandardCharsets.UTF_8);
-    }
-
-    private static String fileExtension(Path file) {
-        String ext = FilenameUtils.getExtension(file.getFileName().toString());
-        return ext == null ? "" : ext.toLowerCase();
-    }
-
-    /** Initializes per-submission lists and an empty composite SBOLDocument. */
-    private void setupReadSbol(SubmitPayload payload) {
-        payload.setAttachmentFiles(new ArrayList<>());
-        payload.setSbolFiles(new ArrayList<>());
-        payload.setUrisFoundInSynBioHub(new HashSet<>());
-
-        SBOLDocument doc = new SBOLDocument();
-        String uriPrefix = collectionService.uriPrefix(payload);
-        if (uriPrefix != null) {
-            doc.setDefaultURIprefix(uriPrefix);
-        }
-        payload.setSbolDocument(doc);
-        payload.setReadSbolStartedAtMs(System.currentTimeMillis());
-        log.debug("readSbol setup: uriPrefix={}", uriPrefix);
     }
 
     // -------------------------------------------------------------------------
@@ -822,7 +160,7 @@ public class SubmitServiceImpl implements SubmitService {
     private void deleteStaggered(String update, String graphUri) throws IOException {
         while (true) {
             String raw = sparqlAuthUpdate(update, graphUri, true);
-            JsonNode bindings = MAPPER.readTree(raw).path("results").path("bindings");
+            JsonNode bindings = objectMapper.readTree(raw).path("results").path("bindings");
             if (!bindings.isArray() || bindings.isEmpty()) {
                 break;
             }
@@ -1018,7 +356,7 @@ public class SubmitServiceImpl implements SubmitService {
                 .loadTemplate(Map.of("uri", collectionUri));
         String raw = searchService.SPARQLQuery(query, graphUri);
         Map<String, String> sources = new HashMap<>();
-        JsonNode bindings = MAPPER.readTree(raw).path("results").path("bindings");
+        JsonNode bindings = objectMapper.readTree(raw).path("results").path("bindings");
         if (!bindings.isArray()) {
             return sources;
         }
@@ -1051,12 +389,12 @@ public class SubmitServiceImpl implements SubmitService {
         templateParams.put("attachmentURI", attachmentUri);
         templateParams.put("attachmentSource", attachmentUri + "/download");
         templateParams.put("persistentIdentity", persistentIdentity);
-        templateParams.put("displayId", sparqlStringLiteral(displayId));
-        templateParams.put("version", sparqlStringLiteral(version));
-        templateParams.put("name", sparqlStringLiteral(name));
-        templateParams.put("description", sparqlStringLiteral(""));
-        templateParams.put("hash", sparqlStringLiteral(uploadHash));
-        templateParams.put("size", sparqlStringLiteral(Long.toString(size)));
+        templateParams.put("displayId", StringUtil.sparqlStringLiteral(displayId));
+        templateParams.put("version", StringUtil.sparqlStringLiteral(version));
+        templateParams.put("name", StringUtil.sparqlStringLiteral(name));
+        templateParams.put("description", StringUtil.sparqlStringLiteral(""));
+        templateParams.put("hash", StringUtil.sparqlStringLiteral(uploadHash));
+        templateParams.put("size", StringUtil.sparqlStringLiteral(Long.toString(size)));
         templateParams.put("type", attachmentType);
         templateParams.put("ownedBy", ownedBy);
         String update = new SPARQLQuery(ATTACH_UPLOAD_SPARQL).loadTemplate(templateParams);
@@ -1070,13 +408,13 @@ public class SubmitServiceImpl implements SubmitService {
         String update = new SPARQLQuery(UPDATE_ATTACHMENT_SPARQL).loadTemplate(Map.of(
                 "attachmentURI", attachmentUri,
                 "attachmentSource", attachmentUri + "/download",
-                "hash", sparqlStringLiteral(uploadHash),
-                "size", sparqlStringLiteral(Long.toString(size))));
+                "hash", StringUtil.sparqlStringLiteral(uploadHash),
+                "size", StringUtil.sparqlStringLiteral(Long.toString(size))));
         sparqlAuthUpdate(update, graphUri, false);
     }
 
     private String attachmentTypeFromExtension(Path file) throws IOException {
-        String ext = fileExtension(file);
+        String ext = FileUtil.fileExtension(file);
         JsonNode mapping = ConfigUtil.get("fileExtensionToAttachmentType");
         if (mapping != null && mapping.has(ext)) {
             return mapping.get(ext).asText();
@@ -1092,7 +430,7 @@ public class SubmitServiceImpl implements SubmitService {
         byte[] raw = Files.readAllBytes(file);
         try {
             MessageDigest digest = MessageDigest.getInstance("SHA-1");
-            String hash = sha1Hex(digest.digest(raw));
+            String hash = sbolService.sha1Hex(digest.digest(raw));
             Path dest = uploadPath(hash);
             if (!Files.exists(dest)) {
                 Files.createDirectories(dest.getParent());
@@ -1108,11 +446,6 @@ public class SubmitServiceImpl implements SubmitService {
 
     private static Path uploadPath(String hash) {
         return Path.of("uploads", hash.substring(0, 2), hash.substring(2) + ".gz");
-    }
-
-    /** Escapes a value for substitution into SPARQL string literal positions in templates. */
-    private static String sparqlStringLiteral(String value) {
-        return "\"" + value.replace("\\", "\\\\").replace("\"", "\\\"") + "\"";
     }
 
     /** Removes unpack directory and prepared SBOL XML after successful upload. */
@@ -1154,14 +487,6 @@ public class SubmitServiceImpl implements SubmitService {
         return ResponseEntity.ok()
                 .contentType(MediaType.parseMediaType("text/plain; charset=UTF-8"))
                 .body("Submission successful");
-    }
-
-    private static String trimToNull(String value) {
-        if (value == null) {
-            return null;
-        }
-        String trimmed = value.trim();
-        return trimmed.isEmpty() ? null : trimmed;
     }
 
     private static String sanitizeFilename(String name) {
